Remove all checks that compare the plaintext being split with the generated ciphertext.
I was previously comparing the generated ciphertext to the plaintext input that was being split.
If any repeating characters were found, I would re-generate a random key, XOR the random key with
the plaintext input, and compare the resulting ciphertext with the plaintext input again.  This
process would repeat until there were no repeating sequences in common being the plaintext and
ciphertext.  While not intentional, I believe this introduced some dependence between the
plaintext input and the randomly generated key, which violates one of the rules of the cryptographic
process for a one-time pad operation:

> The key must be random (uniformly distributed in the set of all possible keys and independent of the plaintext),
> entirely sampled from a non-algorithmic, chaotic source such as a hardware random number generator.

By removing this check, I ensure that there isn't latent dependence introduced between the plaintext
and random key.

The drawback to removing these checks, and reason why they were initially added, is that if the
random key used in the XOR operation with the plaintext is shorter, repeating characters at the
beginning of the ciphertext will match the characters of the plaintext in hexadecimal format.
Extra care must be taken to ensure that the random key is always at least the same length or
longer than the plaintext input after it is converted to hex.

Signed-off-by: Frank Hinek <frankhinek@users.noreply.github.com>